BaseSetting: Data EX
데이터의 위치가 잘못되어 수정. (Data.Contents, Savedata.Contents)
