+++ Library project/Assets/Scripts/Data/Savedata.Contents.cs
@@ -1,39 +1,43 @@
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using System;
 
-//DataManager.cs에서 예시 및 테스트로 사용되는 클래스
+//JsonEx.cs, DataManager에서 예시 및 테스트로 사용되는 클래스
 //흔한 변수명이라 나중에 겹칠까봐 M_을 붙임
-#region  M_Dialogue
-
-public class M_Dialogue{
-    string _name;
-    string _content;
-    int _love;
-
-    public M_Dialogue(string name, string content, int love){
-        _name = name;
-        _content = content;
-        _love = love;
-    }
-
-    public void printName()
-    {
-        Debug.Log("CsvName : " + _name);
-    }
-
-}
-
-public class M_DialogueData : CsvLoader<M_Dialogue>
+#region M_User
+[Serializable]
+public class M_User
 {
-    public List<M_Dialogue> List
+    //private 로 선언해야 할 시 [Serializable Field] 붙여야 함.
+    //대부분의 상황에서는 private 로 선언하는 게 맞음.
+    
+    public int userId;
+    public string firstName;
+
+    public M_User(int id, string name)
     {
-        get => _list;
-        set => _list = value; 
+        userId = id;
+        firstName = name;
     }
-    public override M_Dialogue MakeInstance(string[] parse)
+
+    public void PrintName()
     {
-        return new M_Dialogue(parse[0], parse[1], int.Parse(parse[2]));
+        Debug.Log("printNam" +
+                  "name : " + firstName);
+    }
+}
+[Serializable]
+public class UserData : IJsonLoader<int, M_User>
+{
+    public List<M_User> users = new List<M_User>();
+    
+    public Dictionary<int, M_User> MakeDict()
+    {
+        Dictionary<int, M_User> dict = new Dictionary<int, M_User>();
+        foreach (M_User user in users)
+            dict.Add(user.userId, user);
+        return dict;
     }
 }
